Move l10n & $t mixin to separate directory.
