+++ src/core/l10n/module.js
