+++ src/core/l10n/t.js
