+++ src/core/l10n/translations/de.js
