+++ src/core/l10n/translations/pt.js
